backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
langchain
langchain-community
langchain-openai
langchain-text-splitters
openai==1.81.0
faiss-cpu>=1.7.4
pypdf>=3.15.1
tiktoken>=0.5.1
azure-search-documents>=11.5.2,<12.0.0
